# Conflicts:
#	src/main/java/it/poliba/is45/ecoshipping/controller/OrdineController.java

+++ src/main/java/it/poliba/is45/ecoshipping/domain/Utente.java
@@ -21,7 +21,7 @@ public class Utente {
     @GeneratedValue(strategy = GenerationType.AUTO)
 
     @Column(name = "id_utente")
-    private long idUtente;
+    private int idUtente;
     private String nome;
     private String cognome;
     @Column(name = "indirizzo_residenza")
